:recycle: split proposals filters and actions in 2 components
