+++ src/screens/organizer/proposals/proposalsFilters/index.js
@@ -1,0 +1,1 @@
+export { default } from './proposalsFilters.container'
