+++ src/screens/organizer/proposals/proposalsHeader/proposalsHeader.module.css
@@ -3,7 +3,7 @@
   align-items: center;
   justify-content: space-between;
   width: 100%;
-  margin-bottom: 1em;
+  margin-bottom: 2em;
 }
 
 .title {
