+++ src/screens/organizer/proposals/proposalsToolbar/proposalsToolbar.jsx
@@ -1,7 +1,6 @@
 import React, { Component, Fragment } from 'react'
 import PropTypes from 'prop-types'
 import cn from 'classnames'
-import debounce from 'lodash/debounce'
 import isEmpty from 'lodash/isEmpty'
 
 import Checkbox from 'components/form/checkbox'
@@ -9,39 +8,9 @@ import Button from 'components/button'
 import IconLabel from 'components/iconLabel'
 import styles from './proposalsToolbar.module.css'
 
-const sortOrderLabel = sortOrder => ({
-  newest: 'Newest',
-  oldest: 'Oldest',
-  highestRating: 'Highest Ratings',
-  lowestRating: 'Lowest Ratings',
-}[sortOrder])
-
-const ratingsLabel = rating => ({
-  rated: 'Rated',
-  notRated: 'Not rated',
-}[rating])
-
-const statusLabel = status => ({
-  submitted: 'Not deliberated',
-  accepted: 'Accepted',
-  rejected: 'Rejected',
-  confirmed: 'Confirmed',
-  declined: 'Declined',
-}[status])
-
 class ProposalToolbar extends Component {
-  constructor(props) {
-    super(props)
-    this.onChange = debounce(this.props.onChange, 200)
-  }
-
   state = {
     checkAll: this.props.selection.length,
-  }
-
-  debounceOnChange = (e) => {
-    e.persist()
-    this.onChange(e)
   }
 
   handleSelect = () => {
@@ -53,13 +22,6 @@ class ProposalToolbar extends Component {
 
   render() {
     const {
-      statuses,
-      ratings,
-      formats,
-      categories,
-      sortOrders,
-      filters,
-      onChange,
       onSendEmails,
       onExportProposals,
       onAcceptProposals,
@@ -74,107 +36,49 @@ class ProposalToolbar extends Component {
     const { checkAll } = this.state
 
     return (
-      <div className={cn(styles.proposalsToolbar, 'no-print')}>
-        <div className={styles.proposalsFilters}>
-          <input
-            id="search"
-            type="search"
-            placeholder="Search by title or speaker"
-            onChange={this.debounceOnChange}
-            defaultValue={filters.search}
+      <div className={cn(styles.proposalsActions, 'no-print')}>
+        <div className={styles.leftActions}>
+          <Checkbox
+            onClick={this.handleSelect}
+            label={!nbSelected ? `${totalProposals} proposals` : `${nbSelected} selected`}
+            name="all-pages"
+            value={checkAll}
           />
-
-          {deliberationActive && (
-            <select id="state" onChange={onChange} defaultValue={filters.state}>
-              <option value="">All statuses</option>
-              {statuses.map(status => (
-                <option key={status} value={status}>
-                  {statusLabel(status)}
-                </option>
-              ))}
-            </select>
-          )}
-
-          <select id="ratings" onChange={onChange} defaultValue={filters.ratings}>
-            <option value="">All ratings</option>
-            {ratings.map(rating => (
-              <option key={rating} value={rating}>
-                {ratingsLabel(rating)}
-              </option>
-            ))}
-          </select>
-
-          <select id="formats" onChange={onChange} defaultValue={filters.formats}>
-            <option value="">All formats</option>
-            {formats.map(({ id, name }) => (
-              <option key={id} value={id}>
-                {name}
-              </option>
-            ))}
-          </select>
-
-          <select id="categories" onChange={onChange} defaultValue={filters.categories}>
-            <option value="">All categories</option>
-            {categories.map(({ id, name }) => (
-              <option key={id} value={id}>
-                {name}
-              </option>
-            ))}
-          </select>
-
-          <select id="sortOrder" onChange={onChange} defaultValue={filters.sortOrder}>
-            <option value="">Sort</option>
-            {sortOrders.map(sortOrder => (
-              <option key={sortOrder} value={sortOrder}>
-                {sortOrderLabel(sortOrder)}
-              </option>
-            ))}
-          </select>
         </div>
-        <div className={styles.proposalsActions}>
-          <div className={styles.leftActions}>
-            <Checkbox
-              onClick={this.handleSelect}
-              label={!nbSelected ? `${totalProposals} proposals` : `${nbSelected} selected`}
-              name="all-pages"
-              value={checkAll}
-            />
-          </div>
-          <div className={styles.rightActions}>
-            {nbSelected === 0 && (
-              <Button onClick={onExportProposals} tertiary disabled={isExporting}>
-                <IconLabel
-                  icon="fa fa-cloud-download"
-                  label={isExporting ? 'Exporting...' : 'Export to JSON'}
-                />
+        <div className={styles.rightActions}>
+          {nbSelected === 0 && (
+            <Button onClick={onExportProposals} tertiary disabled={isExporting}>
+              <IconLabel
+                icon="fa fa-cloud-download"
+                label={isExporting ? 'Exporting...' : 'Export to JSON'}
+              />
+            </Button>
+          )}
+          {deliberationActive && nbSelected > 0 && (
+            <Fragment>
+              <Button
+                tertiary
+                onClick={() => onAcceptProposals(selection)}
+                disabled={isEmpty(selection)}
+              >
+                <IconLabel icon="fa fa-check" label="Accept proposals" />
               </Button>
-            )}
-            {deliberationActive && nbSelected > 0 && (
-              <Fragment>
-                <Button
-                  tertiary
-                  onClick={() => onAcceptProposals(selection)}
-                  disabled={isEmpty(selection)}
-                >
-                  <IconLabel icon="fa fa-check" label="Accept proposals" />
-                </Button>
-                <Button
-                  tertiary
-                  onClick={() => onRejectProposals(selection)}
-                  disabled={isEmpty(selection)}
-                >
-                  <IconLabel icon="fa fa-close" label="Reject proposals" />
-                </Button>
-                <Button
-                  tertiary
-                  onClick={() => onSendEmails(selection)}
-                  disabled={isEmpty(selection)}
-                >
-                  <IconLabel icon="fa fa-rocket" label="Send emails" />
-                </Button>
-              </Fragment>
-            )}
-          </div>
+              <Button
+                tertiary
+                onClick={() => onRejectProposals(selection)}
+                disabled={isEmpty(selection)}
+              >
+                <IconLabel icon="fa fa-close" label="Reject proposals" />
+              </Button>
+              <Button
+                tertiary
+                onClick={() => onSendEmails(selection)}
+                disabled={isEmpty(selection)}
+              >
+                <IconLabel icon="fa fa-rocket" label="Send emails" />
+              </Button>
+            </Fragment>
+          )}
         </div>
       </div>
     )
@@ -182,14 +86,7 @@ class ProposalToolbar extends Component {
 }
 
 ProposalToolbar.propTypes = {
-  statuses: PropTypes.arrayOf(PropTypes.string),
   selection: PropTypes.arrayOf(PropTypes.string),
-  ratings: PropTypes.arrayOf(PropTypes.string),
-  formats: PropTypes.arrayOf(PropTypes.object),
-  categories: PropTypes.arrayOf(PropTypes.object),
-  sortOrders: PropTypes.arrayOf(PropTypes.string),
-  filters: PropTypes.objectOf(PropTypes.string),
-  onChange: PropTypes.func.isRequired,
   onSendEmails: PropTypes.func.isRequired,
   onSelectAll: PropTypes.func.isRequired,
   onAcceptProposals: PropTypes.func.isRequired,
@@ -202,13 +99,7 @@ ProposalToolbar.propTypes = {
 }
 
 ProposalToolbar.defaultProps = {
-  statuses: [],
   selection: [],
-  ratings: [],
-  formats: [],
-  categories: [],
-  sortOrders: [],
-  filters: {},
   deliberationActive: false,
   isExporting: false,
   nbSelected: 0,
