+++ src/screens/organizer/proposals/proposalsToolbar/proposalsToolbar.module.css
@@ -1,39 +1,13 @@
-.proposalsToolbar {
-  width: 100%;
-}
-
-.proposalsFilters {
-  display: flex;
-  align-items: center;
-  flex-wrap: wrap;
-  flex-grow: 1;
-  margin-bottom: 1em;
-}
-
-.proposalsFilters > select {
-  width: initial;
-  min-width: 10em;
-  max-width: 15em;
-  margin-left: 1em;
-}
-
-.proposalsFilters > input[type='search'] {
-  width: initial;
-  min-width: 10em;
-  width: initial;
-  flex-grow: 1;
-}
-
 .proposalsActions {
   display: flex;
   align-items: center;
   background: #fff;
   border: 1px solid var(--grey-color);
-  border-top-left-radius: 3px;
-  border-top-right-radius: 3px;
+  border-radius: 3px;
   width: 100%;
   padding: 0.5em 0;
   height: 4em;
+  margin-bottom: 1em;
 }
 
 .leftActions {
@@ -49,8 +23,8 @@
   margin-right: 1em;
 }
 
-@media only screen and (max-width: 1024px) {
-  .proposalsFilters > select {
+@media only screen and (max-width: 480px) {
+  .proposalsActions {
     display: none;
   }
 }
